first attempt to save stuff to public gcs
