+++ _shared_utils/shared_utils/gtfs_analytics_data.yml
@@ -61,6 +61,8 @@ digest_tables:
   operator_profiles: "digest/operator_profiles"
   operator_routes_map: "digest/operator_routes"
   operator_sched_rt: "digest/operator_schedule_rt_category"
+  scheduled_service_hours: "digest/total_scheduled_service_hours"
+  operator_profile_portfolio_view: "digest/operator_profile_portfolio_view"
 
 stop_segments:
   dir: ${gcs_paths.SEGMENT_GCS}
