+++ gtfs_digest/_section1_utils.py
@@ -10,6 +10,7 @@ from segment_speed_utils.project_vars import (COMPILED_CACHED_VIEWS, RT_SCHED_GC
 from shared_utils import catalog_utils, rt_dates, rt_utils
 import _report_utils 
 import _operators_prep as op_prep
+from segment_speed_utils import helpers
 
 # Readable Dictionary
 import yaml
@@ -25,14 +26,6 @@ from IPython.display import HTML, Markdown, display
 with open("color_palettes.yml") as f:
     color_dict = yaml.safe_load(f)
 
-### Delete after incorporating stuff into pipeline
-from segment_speed_utils import helpers, time_series_utils
-
-import os
-from calitp_data_analysis.sql import query_sql
-from calitp_data_analysis.tables import tbls
-from siuba import *
-
 """
 Data
 """
@@ -41,53 +34,14 @@ def organization_name_crosswalk(organization_name: str) -> str:
     Used to match organization_name field with name.
     """
     schd_vp_url = f"{GTFS_DATA_DICT.digest_tables.dir}{GTFS_DATA_DICT.digest_tables.route_schedule_vp}.parquet"
-    og = pd.read_parquet(
+    df = pd.read_parquet(
         schd_vp_url, filters=[[("organization_name", "==", organization_name)]],
     )
     
     # Get most recent name
-    og = og.sort_values(by = ['service_date'], ascending = False)
-    name = og.name.values[0]
+    df = df.sort_values(by = ['service_date'], ascending = False)
+    name = df.name.values[0]
     return name
-
-def load_operator_profiles(organization_name:str)->pd.DataFrame:
-    """
-    Load operator profile dataset for one operator
-    """
-    op_profiles_url = f"{GTFS_DATA_DICT.digest_tables.dir}{GTFS_DATA_DICT.digest_tables.operator_profiles}.parquet"
-
-    op_profiles_df = pd.read_parquet(
-    op_profiles_url,
-    filters=[[("organization_name", "==", organization_name)]])
-    
-    # Keep only the most recent row
-    op_profiles_df1 = op_profiles_df.sort_values(by = ['service_date'], ascending = False).head(1)
-    
-    ntd_cols = [
-        "schedule_gtfs_dataset_key",
-        "counties_served",
-        "service_area_sq_miles",
-        "hq_city",
-        "uza_name",
-        "service_area_pop",
-        "organization_type",
-        "primary_uza",
-        "reporter_type"
-    ]
-    
-    # Load NTD through the crosswalk for the most recent date
-    most_recent_date = rt_dates.y2024_dates[-1]
-    
-    ntd_df = helpers.import_schedule_gtfs_key_organization_crosswalk(most_recent_date)[
-    ntd_cols]
-    
-    # Try to merge
-    op_profiles_df1 = pd.merge(op_profiles_df1, ntd_df, on = ["schedule_gtfs_dataset_key"], how = "left")
-        
-    # Rename dataframe
-    op_profiles_df1.columns = op_profiles_df1.columns.map(_report_utils.replace_column_names)
-    
-    return op_profiles_df1 
 
 def load_operator_map(name:str)->gpd.GeoDataFrame:
     """
@@ -114,6 +68,7 @@ def get_counties():
     my_gdf = to_snakecase(gpd.read_file(f"{ca_gdf}"))[["county_name", "geometry"]]
 
     return my_gdf
+
 """
 Data Manipulation
 Change dataframes from long to wide
@@ -239,7 +194,7 @@ def create_service_hour_chart(df:pd.DataFrame,
                title = _report_utils.labeling(y_col)),
         color=alt.Color(
             "Month",
-            scale=alt.Scale(range=color_dict["four_color"]),  # Specify desired order
+            scale=alt.Scale(range=color_dict["full_color_scale"]),  # Specify desired order
         ),
         opacity=alt.condition(selection, alt.value(1), alt.value(0.2)),
         tooltip=list(df.columns),
@@ -331,126 +286,3 @@ def plot_route(route):
     style_kwds={"weight": 3},
     legend=False,
     tooltip=["Route", "Service Miles"]))
-    
-"""
-Stuff to delete after incorporating into the pipeline 
-"""
-def ntd_operator_info(year:int, organization_name:str)->pd.DataFrame:
-    ntd_mobility_df = merge_ntd_mobility(year)
-    op_profiles = op_prep.operators_with_rt()[['organization_name']]
-    op_profiles = op_profiles.loc[op_profiles.organization_name == organization_name].reset_index(drop = True)
-    m1 = pd.merge(op_profiles, ntd_mobility_df,
-                 how = "inner", left_on = ["organization_name"],
-                 right_on = ["agency_name"])
-    
-    m1 = m1.fillna('None')
-    return m1
-
-def concatenate_trips(
-    date_list: list,
-) -> pd.DataFrame:
-    """
-    Concatenate schedule data that's been
-    aggregated to route-direction-time_period for
-    multiple days.
-    """
-    FILE = GTFS_DATA_DICT.schedule_downloads.trips
-
-    df = (
-        time_series_utils.concatenate_datasets_across_dates(
-            COMPILED_CACHED_VIEWS,
-            FILE,
-            date_list,
-            data_type="df",
-            columns=[
-                "name",
-                "service_date",
-                "route_long_name",
-                "trip_first_departure_datetime_pacific",
-                "service_hours",
-            ],
-        )
-        .sort_values(["service_date"])
-        .reset_index(drop=True)
-    )
-
-    return df
-
-def get_day_type(date):
-    """
-    Function to return the day type (e.g., Monday, Tuesday, etc.) from a datetime object.
-    """
-    days_of_week = ["Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday", "Sunday"]
-    return days_of_week[date.weekday()]
-
-def weekday_or_weekend(row):
-    """
-    Tag if a day is a weekday or Saturday/Sunday
-    """
-    if row.day_type == "Sunday":
-        return "Sunday"
-    if row.day_type == "Saturday":
-        return "Saturday"
-    else:
-        return "Weekday"
-
-def total_service_hours(date_list: list, name: str) -> pd.DataFrame:
-    """
-    Total up service hours by departure hour, 
-    month, and day type for an operator. 
-    """
-    # Combine all the days' data for a week
-    df = concatenate_trips(date_list)
-    
-     # Filter
-    df = df.loc[df.name == name].reset_index(drop=True)
-    
-    # Add day type aka Monday, Tuesday, Wednesday...
-    df['day_type'] = df['service_date'].apply(get_day_type)
-    
-    # Tag if the day is a weekday, Saturday, or Sunday
-    df["weekend_weekday"] = df.apply(weekday_or_weekend, axis=1)
-    
-    # Find the minimum departure hour
-    df["departure_hour"] = df.trip_first_departure_datetime_pacific.dt.hour
-    
-    # Delete out the specific day, leave only month & year
-    df["month"] = df.service_date.astype(str).str.slice(stop=7)
-    
-    df2 = (
-        df.groupby(["name", "month", "weekend_weekday", "departure_hour"])
-        .agg(
-            {
-                "service_hours": "sum",
-            }
-        )
-        .reset_index()
-    )
-    df2["weekday_service_hours"] = df2.service_hours/5
-    df2 = df2.rename(columns = {'service_hours':'weekend_service_hours'})
-    return df2
-
-def total_service_hours_all_months(name: str) -> pd.DataFrame:
-    """
-    Find service hours for a full week for one operator
-    and for the months we have a full week's worth of data downloaded.
-    As of 5/2024, we have April 2023 and October 2023.
-    """
-    # Grab the dataframes with a full week's worth of data. 
-    apr_23week = rt_dates.get_week(month="apr2023", exclude_wed=False)
-    oct_23week = rt_dates.get_week(month="oct2023", exclude_wed=False)
-    apr_24week = rt_dates.get_week(month="apr2024", exclude_wed=False)
-    # need to add april 2024 here 
-    
-    # Sum up total service_hours
-    apr_23df = total_service_hours(apr_23week, name)
-    oct_23df = total_service_hours(oct_23week, name)
-    apr_24df = total_service_hours(apr_24week, name)
-    
-    # Combine everything
-    all_df = pd.concat([apr_23df, oct_23df, apr_24df])
-    
-    # Rename the columns
-    all_df.columns = all_df.columns.map(_report_utils.replace_column_names)
-    return all_df
-
